+++ app/src/main/java/woowacourse/shopping/ui/mapper/CartItemMapper.kt
@@ -1,0 +1,14 @@
+package woowacourse.shopping.ui.mapper
+
+import woowacourse.shopping.domain.model.CartItem
+import woowacourse.shopping.ui.model.CartItemUiModel
+
+object CartItemMapper {
+    fun CartItem.toUiModel() =
+        CartItemUiModel(
+            id = id,
+            quantity = quantity,
+            product = product,
+            checked = false,
+        )
+}
